==> notebooks/1-load-and-convert-metrica-data.ipynb <==
## Metrica Data Processing

- 연구에 활용할 수 있도록 metrica-data전처리하는 코드
- Metrica 1,2,3경기를 loading & preprocess하는 코드

import os
import sys

sys.path.append(os.path.abspath(os.path.dirname(os.getcwd())))

%load_ext autoreload
%autoreload 2

import json
import xml.etree.ElementTree as ET

import numpy as np
import pandas as pd
import torch
from matplotlib import animation
from tqdm import tqdm

from datatools.metrica_helper import MetricaHelper
from datatools.trace_helper import TraceHelper

match_id = 1
team1_file = f"../data/sample-data/Sample_Game_{match_id}/Sample_Game_{match_id}_RawTrackingData_Home_Team.csv"
team2_file = f"../data/sample-data/Sample_Game_{match_id}/Sample_Game_{match_id}_RawTrackingData_Away_Team.csv"
team1_traces = pd.read_csv(team1_file, header=[0, 1, 2])
team2_traces = pd.read_csv(team2_file, header=[0, 1, 2])

team2_traces

### Parsing Metrica Sample Game 3 Data

tree = ET.parse("../data/sample-data/Sample_Game_3/Sample_Game_3_metadata.xml")
root = tree.getroot()
root[0].tag, root[1].tag

player_records = []

for player in root.iter("Player"):
    team_code = player.get("teamId")[-1]
    squad_num = int(player.findtext("ShirtNumber"))
    player_code = f"{team_code}{squad_num:02d}"

    for param in player.iter("ProviderParameter"):
        if param.findtext("Name") == "position_type":
            position = param.findtext("Value")

    player_records.append([squad_num, player_code, position])

player_records = pd.DataFrame(player_records, columns=["squad_num", "code", "position"]).set_index("squad_num")
player_records

phase_records = []

for i, data_spec in enumerate(root[1]):
    start_frame = int(data_spec.get("startFrame"))
    end_frame = int(data_spec.get("endFrame"))
    session = 1 if i == 0 else 2

    player_codes = []
    gk_codes = []

    for player_xy in data_spec[1]:
        squad_num = int(player_xy[0].get("playerChannelId")[6:-2])
        player_code = player_records.at[squad_num, "code"]
        player_codes.append(player_code)

        position = player_records.at[squad_num, "position"]
        if position == "Goalkeeper":
            gk_codes.append(player_code)
    
    player_codes = player_codes[10:11] + player_codes[:10] + player_codes[-1:] + player_codes[11:-1]
    phase_records.append([i + 1, session, start_frame, end_frame, player_codes, gk_codes])

header = ["phase", "session", "start_frame", "end_frame", "player_codes", "gk_codes"]
phase_records = pd.DataFrame(phase_records, columns=header).set_index("phase")
phase_records

### Metrica-data3경기는 추가적인 전처리작업 수행

- metrica-data3의 event데이터는 json형태로 되어있으므로 json -> csv변환
- metrica-data3의 tracking데이터는 txt형태로 되어있으므로 txt -> csv변환

event3_json_file = f"../data/sample-data/Sample_Game_3/Sample_Game_3_events.json"
with open(event3_json_file) as f:
    js = json.loads(f.read()) ## json 라이브러리 이용

# 먼저 json_normalize를 사용하여 "data" 부분을 평탄화합니다.
events3 = pd.json_normalize(js, record_path='data')

# "metadata" 부분을 따로 처리할 필요가 있을 수도 있습니다.
metadata_df = pd.json_normalize(js, record_path='metadata')

rename_columns = {'team.name':'team', 'type.name':'type', 'subtypes.name':'subtype', 'period':'session', 'start.frame':'start_frame', 'start.time':'start_time',
                  'end.frame':'end_frame', 'end.time':'end_time', 'from.name':'from', 'to.name':'to',
                  'start.x':'start_x', 'start.y':'start_y','end.x':'end_x', 'end.y':'end_y'}

events3 = events3[list(rename_columns.keys())]
events3.rename(columns=rename_columns, inplace=True)
events3 = events3[list(rename_columns.values())]
events3['team'] = events3['team'].map({'Team A':'Away', 'Team B':'Home'})

events3.to_csv(f"../data/sample-data/Sample_Game_3/Sample_Game_3_RawEventsData.csv", index=False)
events3

time_cols = ["frame", "session", "time"]
xy_cols = np.array([[f"{p}_x", f"{p}_y"] for p in player_records["code"].tolist() + ["ball"]]).flatten().tolist()

traces_txt = pd.read_csv("../data/sample-data/Sample_Game_3/Sample_Game_3_tracking.txt", sep=";", header=None)
traces = pd.DataFrame(index=traces_txt.index, columns=time_cols + xy_cols)

for phase in tqdm(phase_records.index):
    i0 = phase_records.at[phase, "start_frame"] - 1
    i1 = phase_records.at[phase, "end_frame"] - 1
    player_codes = phase_records.at[phase, "player_codes"]

    phase_traces = traces_txt.loc[i0:i1]
    phase_traces.columns = player_codes
    leftmost = phase_traces[player_codes[0]].str.split(":", expand=True)
    leftmost.columns = ["frame", player_codes[0]]
    rightmost = phase_traces[player_codes[-1]].str.split(":", expand=True)
    rightmost.columns = [player_codes[-1], "ball"]
    phase_traces = pd.concat([leftmost, phase_traces[player_codes[1:-1]], rightmost], axis=1)

    traces.loc[phase_traces.index, "frame"] = phase_traces["frame"].astype(int)
    traces.loc[phase_traces.index, "session"] = phase_records.at[phase, "session"]

    for p in phase_traces.columns[1:]:
        xy = phase_traces[p].str.split(",", expand=True).astype(float).values
        traces.loc[phase_traces.index, [f"{p}_x", f"{p}_y"]] = xy

traces["time"] = (traces["frame"] * 0.04).astype(float).round(2)
traces.to_csv(f"../data/sample-data/Sample_Game_3/Sample_Game_3_RawTrackingData.csv", index=False)
traces

### Processing Metrica Data

- metrica데이터의 여러 코드 작업(MetricaHelper클래스)

1. tracking-data & event-data의 경기장 규격 맞추기
2. 

import warnings
warnings.filterwarnings('ignore')

match_ids = [1, 2, 3]

for match_id in match_ids:

    event_file = f"../data/sample-data/Sample_Game_{match_id}/Sample_Game_{match_id}_RawEventsData.csv"
    events = pd.read_csv(event_file)

    if match_id <= 2:
        team1_file = f"../data/sample-data/Sample_Game_{match_id}/Sample_Game_{match_id}_RawTrackingData_Home_Team.csv"
        team2_file = f"../data/sample-data/Sample_Game_{match_id}/Sample_Game_{match_id}_RawTrackingData_Away_Team.csv"
        team1_traces = pd.read_csv(team1_file, header=[0, 1, 2])
        team2_traces = pd.read_csv(team2_file, header=[0, 1, 2])
        helper = MetricaHelper(team1_traces, team2_traces, events=events)
    else:  # match_id == 3
        trace_file = f"../data/sample-data/Sample_Game_{match_id}/Sample_Game_{match_id}_RawTrackingData.csv"
        traces = pd.read_csv(trace_file, index_col=0)
        helper = MetricaHelper(traces_from_txt=traces, events=events)

    # phase정의 함수 : phase란 선수 교체/퇴장이 발생하거나 후반전이 시작하는 상황을 의미한다.
    # phase가 음수인 상황은 한 팀의 측정된 필드플레이어 수가 10명 미만이 되는 경우로 분석에 사용되지 않는다.
    helper.generate_phase_records()
    helper.generate_phase()

    # 패스의 초기 위치가 부자연스러워지는 문제 때문에 다운샘플링 사용하지 않음 -> helper.downsample_to_10fps() 함수 사용X
    # 자세한 이유: 추적 데이터에는 측정되지 않은 위치값이 많아 보간 필수.
    # 기존 코드는 25fps에서 10fps로 다운샘플링 후 보간 사용. 
    # 그러나 본 연구에서는 0.4초 동안의 패스 위치가 중요한데, 다운샘플링으로 인해 패스 방향이 부자연스러운 장면이 생김
    # helper.downsample_to_10fps()

    # episode정의 함수 : CARD, OFFSIDE등의 상황으로 인해 경기가 중단된 상황을 의미한다.
    helper.split_into_episodes()

    # 공과 선수들의 위치좌표를 바탕으로 SPEED, ACCEL계산
    helper.calc_running_features(remove_outliers=True, smoothing=True)

    # 추적데이터의 소유권 정의 함수
    helper.find_gt_player_poss()
    helper.find_gt_team_poss()

    # tracking_data와 event_data의 player_possession정보가 mismatch되는 문제
    if match_id == 3:
        helper.correct_event_player_ids()

    helper.traces.to_csv(f"../data/preprocess-data/tracking-data/match{match_id}.csv", index=False)
    helper.events.to_csv(f"../data/preprocess-data/event-data/match{match_id}.csv", index=False)

helper.events

helper.traces

col_x = [col for col in helper.traces.columns if col.endswith("_x")]
col_y = [col for col in helper.traces.columns if col.endswith("_y")]

print("x좌표 통계 : ", helper.traces[col_x].max().max(), helper.traces[col_x].min().min())
print("y좌표 통계 : ", helper.traces[col_y].max().max(), helper.traces[col_y].min().min())
