==> notebooks/2-compute-features-and-labels.ipynb <==
### Metrica data unified

- 연구에 활용될 데이터 추출
- 이벤트 유형 통합

import os, sys
import tqdm
sys.path.append(os.path.abspath(os.path.dirname(os.getcwd())))

import numpy as np
import pandas as pd
import ast
import matplotlib.pyplot as plt

from datatools.features import unified_features
from datatools.labels import unified_labels

game_ids = [1, 2, 3]

all_events = []
for game_id in game_ids:
    events = pd.read_csv(f'../data/preprocess-data/event-data/match{game_id}.csv')
    events['game_id'] = game_id
    
    events = unified_features(events, game_id)
    events = unified_labels(events)
    
    events.to_csv(f'../data/unified-data/match{game_id}.csv',index=False)
    
    all_events.append(events)

all_events = pd.concat(all_events).reset_index(drop=True)
all_events['event_id'] = np.arange(len(all_events))
all_events.to_csv('../data/unified-data/all-match.csv',index=False)

all_events[all_events["eventType"] == "PASS"]["outcome"].value_counts()

print(all_events[all_events["eventType"] == "PASS"]["type"].unique())
print(all_events[all_events["eventType"] == "PASS"]["subtype"].unique())

print("\n성공한 패스 : ", all_events[(all_events["eventType"] == "PASS") & (all_events["outcome"])]["type"].unique())
print("성공한 패스 : ", all_events[(all_events["eventType"] == "PASS") & (all_events["outcome"])]["subtype"].unique())

print("\n실패한 패스 : ", all_events[(all_events["eventType"] == "PASS") & (~all_events["outcome"]) & (~all_events["subtype"].isin(["HEAD", "HEAD-CLEARANCE"]))]["type"].unique())
print("실패한 패스 : ", all_events[(all_events["eventType"] == "PASS") & (~all_events["outcome"]) & (~all_events["subtype"].isin(["HEAD", "HEAD-CLEARANCE"]))]["subtype"].unique())

all_events[all_events["eventType"] == "PASS"].head()

print(all_events[all_events["eventType"] == "SHOT"]["type"].unique())
print(all_events[all_events["eventType"] == "SHOT"]["subtype"].unique())

all_events[all_events["eventType"] == "SHOT"].head()

print(all_events[all_events["eventType"] == "DRIBBLE"]["type"].unique())
print(all_events[all_events["eventType"] == "DRIBBLE"]["subtype"].unique())

all_events[all_events["eventType"] == "DRIBBLE"].head()

print(all_events[all_events["eventType"] == "CROSS"]["type"].unique())
print(all_events[all_events["eventType"] == "CROSS"]["subtype"].unique())

all_events[all_events["eventType"] == "CROSS"].head()
