==> notebooks/4-analyze-Intended-receiver.ipynb <==
### 실패한 패스의 Intended-receiver레이블링 작업 통합

- 애니메이션 제작 후 레이블링 수행
- 두명의 연구원의 교차검증을 통해 레이블링 수행

import os, sys
import tqdm
sys.path.append(os.path.abspath(os.path.dirname(os.getcwd())))

import numpy as np
import pandas as pd
import ast
import matplotlib.pyplot as plt

- PassExclude설명

1. (25개): 패스가 아님
2. (3개): 애매한 패스(패스인지 아닌지 불분명)
3. (2개): 데이터 자체의 오류

unsuccess_passes = pd.read_csv('../data/unified-data/unsuccess_passes.csv')
unsuccess_passes

all_events = pd.read_csv("../data/unified-data/all-match.csv")
all_events

all_events[(all_events["eventType"].isin(["PASS","CROSS"])) & (~all_events["outcome"]) & (all_events["start_frame"] < all_events["end_frame"])]

labeling_all_events = pd.merge(all_events, unsuccess_passes, on = "event_id", how="left", suffixes=("","_temp"))
labeling_all_events

col = [col for col in labeling_all_events.columns if not col.endswith("_temp")]
labeling_all_events[col]

labeling_all_events[(labeling_all_events["eventType"].isin(["PASS","CROSS"])) & (~labeling_all_events["outcome"]) & (labeling_all_events["start_frame"] < labeling_all_events["end_frame"])]

labeling_all_events[col].to_csv("../data/unified-data/labeling-all-match.csv", index = False)
